Shifted complex todo repo queries to JPA specifications implementation

+++ grassroot-services/src/main/java/za/org/grassroot/services/async/ScheduledTasks.java
@@ -148,7 +148,7 @@ public class ScheduledTasks {
     @Scheduled(fixedRate = 300000) //runs every 5 minutes
     public void sendTodoReminders() {
         List<Todo> todos = todoRepository.findAllTodosForReminding(Instant.now(),
-                environment.getProperty("grassroot.todos.completion.threshold", Integer.class));
+                environment.getProperty("grassroot.todos.completion.threshold", Double.class));
         logger.info("Sending scheduled reminders for {} todos", todos.size());
 
         for (Todo todo : todos) {
